+++ fatcode/urls.py
@@ -19,7 +19,8 @@ urlpatterns = [
     path('api/v1/dashboard/', include('src.dashboard.urls')),
     path('api/v1/repository/', include('src.repository.urls')),
     path('api/v1/data/', include('src.data.urls')),
-    path('api/v1/support/', include('src.support.urls'))
+    path('api/v1/support/', include('src.support.urls')),
+    path('api/v1/questionnaire/', include('src.questionnaire.urls'))
 ]
 
 urlpatterns += doc_urls
